tic-tac-toe-backend/requirements.txt
Flask
flask_cors
pytest
python-dotenv
psycopg2


